backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
marshmallow==3.22.0
structlog==24.4.0
pytest==8.3.3
pytest-flask==1.3.0
coverage==7.6.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
gunicorn==23.0.0
